requirements.txt
absl-py==2.3.1
addict==2.4.0
basicsr==1.4.2
certifi==2025.7.9
charset-normalizer==3.4.2
contourpy==1.3.2
controlnet_aux==0.0.10
cycler==0.12.1
diffusers==0.34.0
einops==0.8.1
facexlib==0.3.0
filelock==3.18.0
filterpy==1.4.5
fonttools==4.59.0
fsspec==2025.5.1
future==1.0.0
gfpgan==1.3.8
grpcio==1.73.1
hf-xet==1.1.5
huggingface-hub==0.33.4
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
Jinja2==3.1.6
kiwisolver==1.4.8
lazy_loader==0.4
llvmlite==0.44.0
lmdb==1.7.2
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
protobuf==6.31.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
realesrgan==0.3.0
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.16.0
sentencepiece==0.2.0
setuptools==78.1.1
six==1.17.0
sympy==1.14.0
tb-nightly==2.21.0a20250723
tensorboard-data-server==0.7.2
tifffile==2025.6.11
timm==1.0.17
tokenizers==0.21.2
torch==2.7.1
torchsde==0.2.6
torchvision==0.22.1
tqdm==4.67.1
trampoline==0.1.2
transformers==4.53.2
triton==3.3.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
yapf==0.43.0
zipp==3.23.0
